Added Bootstrap, began Navbar
Added Bootstrap to the project and began working on a theme in _content.scss. Also created a version of the logo including the name in Illustrator, for the Navbar. Need to add a button to the Navbar to switch themes and fix the dropdown menu when the screen is too small. Should also look into combining the two logo-name.svg files and changing the color via css.